Change function args from `String` to `Into<String>`
Doesn't help quite as much as one might hope because there's
still lots of structures that you build literally, and those all
contain `String` and kind of have to.  If you want me to make
constructor functions for those I will, but that's an API decision
I happily leave up to you.